+++ src/components/ServiceForm/ServiceForm.module.css
@@ -1,0 +1,4 @@
+.form {
+  width: 100%;
+  margin: 2em 0;
+}
